tunnel {aakmisc}	R Documentation

ssh tunneling

Description

Setting up ssh tunnels for database access.

Usage

startTunnel(
  port = NULL,
  remotehost = getOption("aakmisc.remotehost", NULL),
  user = getOption("aakmisc.user", NULL),
  sleep = 5
)

stopTunnel(..., pid = getOption("aakmisc.tunnelpid", NULL))

Arguments

port
Port on which to connect to PostgreSQL database. If NULL, a random port number will be used.

remotehost
Hostname of PostgreSQL server. An ssh tunnel to this host will be created.

user
Username to use in conneting to PostgreSQL database. If NULL, Sys.getenv("USER") will be used.

sleep
Time in seconds to sleep after initiating the ssh tunnel.

...
Additional arguments will be passed to dbConnect.

pid
ID of ssh tunnel process. Set automatically by startTunnel.

[Package aakmisc version 0.29-1 Index]