trnc {aakmisc}	R Documentation

Truncation of plots.

Description

Truncate to the specified window.

Usage

trnc(x, range = c(0, 1), only.finite = TRUE)

Arguments

x
Numeric vector of values to manipulate.

range
Numeric vector of length two giving desired output range.

only.finite
if TRUE (the default), will only modify finite values.

Details

trnc is a function for truncating data to a specified window. It is suitable for use in scale_{x,y}_{continuous,discrete}, for example.

See Also

censor

Examples

trnc(c(-1,0.5,1,2,NA))

[Package aakmisc version 0.32-2 Index]